# Improve performance of hashing and reduce memory