@Webreaper 2023-08-23
This is totally awesome - thank you! I'll take a look later and merge.

That is a *seriously* good memory optimisation. 😁👌